Add files via upload
In this dataset, the information of 539 women is collected. Part of them has been diagnosed with early-stage breast cancer. In addition, we collect information
on 55 explanatory variables (including binary and continuous variables). We aim to predict the breast cancer status of the patients based on the 55 explanatory variables.

column 1 is the breast cancer status (1-Diagnosed with Breast Cancer,2-Not Diagnosed). Column 2 is the age. Column 3 is the ethnic (1-Hispanic, 0-Not Hispanic). Column 4 is the race (0-White, 1-Black, 2-Asian, 3-Native American). The specifications of the other 52 variables are unavailable. The first 300 observations are training data with breast cancer status, while the remaining 239 observations are the test data without breast cancer status (written as -1 in the Dataset)